Merge Pull Request (#2)
* Edits to File Header for Licence Handling

Changing where the date of the copyright goes (follow the WPILib model).

* Edits to Spacing on package and imports

Need to get an auto-formatter working.

* Update to New WPILib Version

New version of WPILib.

* New Logger Implementation

Hides actual implementation of logger to logger package.

* Remove RPI Module

Not planned for use, and have an example module in the PDB.

* Remove Wheel Color Sensor

Not general and we have example sensors to use.

* Add Auto-update of Build Info from Gradle Build

Matt's new stuff added to build.grade.

* Control of Template Config Files

Configuration files for dashboard and properties. Dashboard goes to the driver station and properties goes to the robot.

* Update Preferences Files

Probably going to refactor which will change this again.

* Refactor of Preferences to Subsystem

Put all the knowledge of the preferences of a subsystem with the subsystem package itself.

* Remove Unused, Non-Generic Stuff

Further trimming of stuff for the framework. Either not used generically or really specific to game that Zester played.

* More Refactor of props and properties

Improving implementation for less required editing on different configurations.

* More Refactor of props and prefs

Including using logger of where getting logged so clear who called.

* Default Impl Class Loading

If the property exists but is not specified, goes with the "real" implementation.  Adds telemetry for what is loaded.

Co-authored-by: Stu Lewin <java4first@thelewins.net>